reorganized testing mocks
