+++ modules/javafx.graphics/src/test/java/test/javafx/scene/command/CommandTest.java
@@ -23,6 +23,8 @@ package test.javafx.scene.command;
 
 import com.sun.javafx.scene.NodeHelper;
 import org.junit.jupiter.api.Test;
+import test.javafx.scene.command.mocks.TestCommand;
+import test.javafx.scene.command.mocks.TestEventBinding;
 import javafx.scene.layout.Pane;
 import java.util.ArrayList;
 import java.util.List;
